otel-microservice-lab/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.1.18
redis>=5.0.4
celery>=5.4.0
prometheus-client>=0.20.0
python-json-logger>=2.0.7
pydantic>=2.7.1
pydantic-settings>=2.2.1
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0
opentelemetry-instrumentation-fastapi>=0.46b0
opentelemetry-instrumentation-psycopg>=0.46b0
opentelemetry-instrumentation-redis>=0.46b0
opentelemetry-instrumentation-celery>=0.46b0
opentelemetry-instrumentation-logging>=0.46b0
